Base Code
So far the code takes a user's input as what they want to yell, and types it out.  What I need to do now is make it more dynamic - maybe the text will shake - and to find a way to clear it after every word, so each one appears individually before clearing out and becoming the next word.